fix: extract random player
